+++ IHMCCommunication/src/us/ihmc/communication/blackoutGenerators/CommunicationBlackoutGenerator.java
@@ -2,5 +2,5 @@ package us.ihmc.communication.blackoutGenerators;
 
 public interface CommunicationBlackoutGenerator
 {
-   public int calculateNextBlackoutLength(int currentTime);
+   public long calculateNextBlackoutLength(long currentTime);
 }
